flask-server/requirements.txt
Flask
Flask-SQLAlchemy
argon2-cffi
argon2
PyJWT
MySQL-python
python-dotenv